Grand Views Refactoring
