+++ lib/src/DateTimePicker/DateTimePicker.tsx
@@ -1,109 +1,63 @@
-import * as React from 'react';
-import DateTimePickerRoot, { BaseDateTimePickerProps } from './DateTimePickerRoot';
 import { useUtils } from '../_shared/hooks/useUtils';
 import { BasePickerProps } from '../typings/BasePicker';
-import { toShowDateTimePickerTabs } from '../_helpers/utils';
-import { ExtendWrapper, Wrapper } from '../wrappers/Wrapper';
-import { usePickerState } from '../_shared/hooks/usePickerState';
+import { BaseDatePickerProps } from '../DatePicker/DatePicker';
+import { BaseTimePickerProps } from '../TimePicker/TimePicker';
+import { DateTimePickerToolbar } from './DateTimePickerToolbar';
+import { pick12hOr24hFormat } from '../_helpers/text-field-helper';
 import { dateTimePickerDefaultProps } from '../constants/prop-types';
-import { PureDateInput, PureDateInputProps } from '../_shared/PureDateInput';
-import { DateValidationProps, validate, pick12hOr24hFormat } from '../_helpers/text-field-helper';
+import { WrappedPurePickerProps, makePurePicker } from '../Picker/WrappedPurePicker';
+import { makeKeyboardPicker, WrappedKeyboardPickerProps } from '../Picker/WrappedKeyboardPicker';
 
-export type DateTimePickerProps = BasePickerProps &
-  DateValidationProps &
-  BaseDateTimePickerProps &
-  ExtendWrapper<PureDateInputProps>;
+export type DateTimePickerView = 'year' | 'date' | 'month' | 'hours' | 'minutes';
 
-export const DateTimePicker: React.FC<DateTimePickerProps> = props => {
-  const {
-    allowKeyboardControl,
-    ampm,
-    animateYearScrolling,
-    autoOk,
-    autoSubmit,
-    dateRangeIcon,
-    disableFuture,
-    disablePast,
-    format,
-    forwardedRef,
-    initialFocusedDate,
-    invalidDateMessage,
-    labelFunc,
-    leftArrowIcon,
-    leftArrowButtonProps,
-    maxDate,
-    maxDateMessage,
-    minDate,
-    minDateMessage,
-    minutesStep,
-    onAccept,
-    onChange,
-    onOpen,
-    onClose,
-    onMonthChange,
-    onYearChange,
-    openTo,
-    renderDay,
-    rightArrowIcon,
-    rightArrowButtonProps,
-    shouldDisableDate,
-    showTabs,
-    timeIcon,
-    value,
-    variant,
-    ...other
-  } = props;
+export type BaseDateTimePickerProps = BaseTimePickerProps & BaseDatePickerProps;
 
+export interface DateTimePickerViewsProps extends BasePickerProps, BaseDateTimePickerProps {
+  /** Array of views to show */
+  views?: ('year' | 'date' | 'month' | 'hours' | 'minutes')[];
+  /** Open to DatePicker */
+  openTo?: 'year' | 'date' | 'month' | 'hours' | 'minutes';
+  /** To show tabs */
+  hideTabs?: boolean;
+  /** Date tab icon */
+  dateRangeIcon?: React.ReactNode;
+  /** Time tab icon */
+  timeIcon?: React.ReactNode;
+}
+
+export type DateTimePickerProps = WrappedPurePickerProps & DateTimePickerViewsProps;
+
+export type KeyboardDateTimePickerProps = WrappedKeyboardPickerProps & DateTimePickerViewsProps;
+
+const defaultProps = {
+  ...dateTimePickerDefaultProps,
+  wider: true,
+  openTo: 'date' as DateTimePickerView,
+  views: ['year', 'date', 'hours', 'minutes'] as DateTimePickerView[],
+};
+
+function useOptions(props: DateTimePickerViewsProps) {
   const utils = useUtils();
-  const toShowTabs = toShowDateTimePickerTabs(showTabs);
-  const { pickerProps, inputProps, wrapperProps } = usePickerState(props, {
-    getValidationError: () => validate(value, utils, props),
+
+  return {
     getDefaultFormat: () =>
-      pick12hOr24hFormat(format, ampm, {
+      pick12hOr24hFormat(props.format, props.ampm, {
         '12h': utils.dateTime12hFormat,
         '24h': utils.dateTime24hFormat,
       }),
-  });
+  };
+}
 
-  return (
-    <Wrapper
-      variant={variant}
-      showTabs={toShowTabs}
-      InputComponent={PureDateInput}
-      DateInputProps={inputProps}
-      {...wrapperProps}
-      {...other}
-    >
-      <DateTimePickerRoot
-        {...pickerProps}
-        ampm={ampm}
-        allowKeyboardControl={allowKeyboardControl}
-        minutesStep={minutesStep}
-        animateYearScrolling={animateYearScrolling}
-        autoSubmit={autoSubmit}
-        dateRangeIcon={dateRangeIcon}
-        disableFuture={disableFuture}
-        disablePast={disablePast}
-        leftArrowIcon={leftArrowIcon}
-        leftArrowButtonProps={leftArrowButtonProps}
-        maxDate={maxDate}
-        minDate={minDate}
-        onMonthChange={onMonthChange}
-        onYearChange={onYearChange}
-        openTo={openTo}
-        renderDay={renderDay}
-        rightArrowIcon={rightArrowIcon}
-        rightArrowButtonProps={rightArrowButtonProps}
-        shouldDisableDate={shouldDisableDate}
-        showTabs={toShowTabs}
-        timeIcon={timeIcon}
-      />
-    </Wrapper>
-  );
-};
+export const DateTimePicker = makePurePicker<DateTimePickerProps>({
+  useOptions,
+  ToolbarComponent: DateTimePickerToolbar,
+});
 
-DateTimePicker.defaultProps = dateTimePickerDefaultProps;
+export const KeyboardDateTimePicker = makeKeyboardPicker<KeyboardDateTimePickerProps>({
+  useOptions,
+  ToolbarComponent: DateTimePickerToolbar,
+});
 
-export default React.forwardRef((props: DateTimePickerProps, ref) => (
-  <DateTimePicker {...props} forwardedRef={ref} />
-));
+DateTimePicker.defaultProps = defaultProps;
+
+KeyboardDateTimePicker.defaultProps = defaultProps;
