--- lib/src/DateTimePicker/components/DateTimePickerView.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react';
-
-export interface DateTimePickerViewProps {
-  selected: boolean;
-  children: React.ReactChild;
-}
-
-export const DateTimePickerView: React.SFC<DateTimePickerViewProps> = ({ selected, children }) => {
-  if (!selected) {
-    return null;
-  }
-
-  return <div children={children} />;
-};
-
-export default DateTimePickerView;
