+++ lib/src/DateTimePicker/index.ts
@@ -1,10 +1,7 @@
-import { DateTimePickerProps } from './DateTimePicker';
-import { KeyboardDateTimePickerProps } from './KeyboardDateTimePicker';
+import { DateTimePickerProps, KeyboardDateTimePickerProps } from './DateTimePicker';
 
 export type DateTimePickerProps = DateTimePickerProps;
 
 export type KeyboardDateTimePickerProps = KeyboardDateTimePickerProps;
 
-export { KeyboardDateTimePicker } from './KeyboardDateTimePicker';
-
-export { default } from './DateTimePicker';
+export { KeyboardDateTimePicker, DateTimePicker } from './DateTimePicker';
