+++ lib/src/__tests__/e2e/DatePickerProps.test.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import DatePicker from '../../DatePicker/DatePicker';
 import { mount, utilsToUse } from '../test-utils';
+import { DatePicker } from '../../DatePicker/DatePicker';
 
 describe('DatePicker - different props', () => {
   it('Should not render toolbar if onlyCalendar = true', () => {
     const component = mount(
       <DatePicker
         open
-        onlyCalendar
+        disableToolbar
         onChange={jest.fn()}
         value={utilsToUse.date('2018-01-01T00:00:00.000Z')}
       />
