+++ lib/src/_helpers/utils.ts
@@ -1,8 +1,3 @@
-export function toShowDateTimePickerTabs(showTabsProps: boolean | undefined) {
-  // do not show tabs for small screens
-  return Boolean(showTabsProps && typeof window !== 'undefined' && window.innerHeight > 667);
-}
-
 export function handleKeydown(
   e: KeyboardEvent,
   keyHandlers: Record<KeyboardEvent['key'], (e: KeyboardEvent) => void>
