--- lib/src/constants/DatePickerView.ts
@@ -1,1 +1,0 @@
-export type DatePickerViewType = 'year' | 'month' | 'day';
